+++ app/views/layouts/application.html.erb
@@ -14,6 +14,5 @@
   <body>
     <%= render 'shared/navbar' %>
     <%= yield %>
-    <%= render 'shared/footer' %>
   </body>
 </html>
