+++ app/views/products/new.html.erb
@@ -1,0 +1,5 @@
+<h1>New Product</h1>
+
+<%= render 'form', product: @product %>
+
+<%= link_to 'Back', products_path %>
